chore: organize code in google translator
